Address PR feedback
Remove FunctionProperties from ExpressionConfig and create it
when create<LANG>Service is called. This ensures it's created for each
query.

Created FunctionPropertiesTestConfig to simplify updating unit tests
to work with the change above.

Removed an infrequently-used version of
BuiltinFunctionRepository.compile and
@Getter on BuiltinFunctionRepository.functionProperties -- it was only
used for a test -- ExpressionConfigTest. Removed the test because
it was testing Spring behavior and not very valuable.

Signed-off-by: MaxKsyunz <maxk@bitquilltech.com>