Add TIME_FORMAT() Function To SQL Plugin
Adds the time_format function to the SQL plugin. The function takes a DATE/DATETIME/TIME/TIMESTAMP/STRING in a valid format, and a formatting string (e.g., "%h %p") and outputs a formatted string. If a DATE is passed in, it treats the argument as if it is taken at midnight (i.e., 00:00:00).

Signed-off-by: GabeFernandez310 <Gabriel.Fernandez@improving.com>